reorganization for begin-set-end functions - no functional changes
